Moved tools out of RenderContext into NotificationRenderContext
