+++ Game/NotificationRenderContext.h
@@ -7,6 +7,7 @@
 
 #include "Font.h"
 #include "GlobalRenderContext.h"
+#include "RenderParameters.h"
 #include "ResourceLocator.h"
 #include "ShaderTypes.h"
 #include "TextureAtlas.h"
@@ -19,6 +20,7 @@
 #include <algorithm>
 #include <array>
 #include <memory>
+#include <optional>
 #include <string>
 #include <vector>
 
@@ -42,30 +44,13 @@ public:
 	NotificationRenderContext(
         ResourceLocator const & resourceLocator,
         ShaderManager<ShaderManagerTraits> & shaderManager,
-		GlobalRenderContext const & globalRenderContext,
-        int canvasWidth,
-        int canvasHeight,
-        float effectiveAmbientLightIntensity);
-
-    void UpdateCanvasSize(int width, int height)
-    {
-		// Recalculate screen -> NDC conversion factors
-        mScreenToNdcX = 2.0f / static_cast<float>(width);
-        mScreenToNdcY = 2.0f / static_cast<float>(height);
-
-		// Make sure we re-calculate (and re-upload) all vertices
-		// at the next iteration
-		for (auto & fc : mFontRenderContexts)
-		{
-			fc.SetLineDataDirty(true);
-		}
-    }
-
-    void UpdateEffectiveAmbientLightIntensity(float effectiveAmbientLightIntensity);
+		GlobalRenderContext const & globalRenderContext);
 
 public:
 
-	void UploadTextNotificationStart(FontType fontType)
+	void UploadStart();
+
+	inline void UploadTextNotificationStart(FontType fontType)
 	{
 		//
 		// Text notifications are sticky: we upload them once in a while and
@@ -78,7 +63,7 @@ public:
 		fontRenderContext.SetLineDataDirty(true);
 	}
 
-	void UploadTextNotificationLine(
+	inline void UploadTextNotificationLine(
 		FontType font,
 		std::string const & text,
 		AnchorPositionType anchor,
@@ -98,12 +83,12 @@ public:
 			alpha);
 	}
 
-	void UploadTextNotificationEnd(FontType /*fontType*/)
+	inline void UploadTextNotificationEnd(FontType /*fontType*/)
 	{
 		// Nop
 	}
 
-	void UploadTextureNotificationStart()
+	inline void UploadTextureNotificationStart()
 	{
 		//
 		// Texture notifications are sticky: we upload them once in a while and
@@ -121,10 +106,114 @@ public:
         vec2f const & screenOffset, // In texture-size fraction (0.0 -> 1.0)
         float alpha);
 
-	void UploadTextureNotificationEnd()
+	inline void UploadTextureNotificationEnd()
 	{
 		// Nop
 	}
+
+	inline void UploadHeatBlasterFlame(
+		vec2f const & centerPosition,
+		float radius,
+		HeatBlasterActionType action)
+	{
+		//
+		// Populate vertices
+		//
+
+		float const quadHalfSize = (radius * 1.5f) / 2.0f; // Add some slack for transparency
+		float const left = centerPosition.x - quadHalfSize;
+		float const right = centerPosition.x + quadHalfSize;
+		float const top = centerPosition.y + quadHalfSize;
+		float const bottom = centerPosition.y - quadHalfSize;
+
+		// Triangle 1
+
+		mHeatBlasterFlameVertexBuffer[0].vertexPosition = vec2f(left, bottom);
+		mHeatBlasterFlameVertexBuffer[0].flameSpacePosition = vec2f(-0.5f, -0.5f);
+
+		mHeatBlasterFlameVertexBuffer[1].vertexPosition = vec2f(left, top);
+		mHeatBlasterFlameVertexBuffer[1].flameSpacePosition = vec2f(-0.5f, 0.5f);
+
+		mHeatBlasterFlameVertexBuffer[2].vertexPosition = vec2f(right, bottom);
+		mHeatBlasterFlameVertexBuffer[2].flameSpacePosition = vec2f(0.5f, -0.5f);
+
+		// Triangle 2
+
+		mHeatBlasterFlameVertexBuffer[3].vertexPosition = vec2f(left, top);
+		mHeatBlasterFlameVertexBuffer[3].flameSpacePosition = vec2f(-0.5f, 0.5f);
+
+		mHeatBlasterFlameVertexBuffer[4].vertexPosition = vec2f(right, bottom);
+		mHeatBlasterFlameVertexBuffer[4].flameSpacePosition = vec2f(0.5f, -0.5f);
+
+		mHeatBlasterFlameVertexBuffer[5].vertexPosition = vec2f(right, top);
+		mHeatBlasterFlameVertexBuffer[5].flameSpacePosition = vec2f(0.5f, 0.5f);
+
+		//
+		// Store shader
+		//
+
+		switch (action)
+		{
+			case HeatBlasterActionType::Cool:
+			{
+				mHeatBlasterFlameShaderToRender = Render::ProgramType::HeatBlasterFlameCool;
+				break;
+			}
+
+			case HeatBlasterActionType::Heat:
+			{
+				mHeatBlasterFlameShaderToRender = Render::ProgramType::HeatBlasterFlameHeat;
+				break;
+			}
+		}
+	}
+
+	inline void UploadFireExtinguisherSpray(
+		vec2f const & centerPosition,
+		float radius)
+	{
+		//
+		// Populate vertices
+		//
+
+		float const quadHalfSize = (radius * 3.5f) / 2.0f; // Add some slack to account for transparency
+		float const left = centerPosition.x - quadHalfSize;
+		float const right = centerPosition.x + quadHalfSize;
+		float const top = centerPosition.y + quadHalfSize;
+		float const bottom = centerPosition.y - quadHalfSize;
+
+		// Triangle 1
+
+		mFireExtinguisherSprayVertexBuffer[0].vertexPosition = vec2f(left, bottom);
+		mFireExtinguisherSprayVertexBuffer[0].spraySpacePosition = vec2f(-0.5f, -0.5f);
+
+		mFireExtinguisherSprayVertexBuffer[1].vertexPosition = vec2f(left, top);
+		mFireExtinguisherSprayVertexBuffer[1].spraySpacePosition = vec2f(-0.5f, 0.5f);
+
+		mFireExtinguisherSprayVertexBuffer[2].vertexPosition = vec2f(right, bottom);
+		mFireExtinguisherSprayVertexBuffer[2].spraySpacePosition = vec2f(0.5f, -0.5f);
+
+		// Triangle 2
+
+		mFireExtinguisherSprayVertexBuffer[3].vertexPosition = vec2f(left, top);
+		mFireExtinguisherSprayVertexBuffer[3].spraySpacePosition = vec2f(-0.5f, 0.5f);
+
+		mFireExtinguisherSprayVertexBuffer[4].vertexPosition = vec2f(right, bottom);
+		mFireExtinguisherSprayVertexBuffer[4].spraySpacePosition = vec2f(0.5f, -0.5f);
+
+		mFireExtinguisherSprayVertexBuffer[5].vertexPosition = vec2f(right, top);
+		mFireExtinguisherSprayVertexBuffer[5].spraySpacePosition = vec2f(0.5f, 0.5f);
+
+		//
+		// Store shader
+		//
+
+		mFireExtinguisherSprayShaderToRender = Render::ProgramType::FireExtinguisherSpray;
+	}
+
+	void UploadEnd();
+
+	void ProcessParameterChanges(RenderParameters const & renderParameters);
 
 	void Draw();
 
@@ -134,9 +223,14 @@ private:
 
 	void GenerateTextVertices(FontRenderContext & context);
 
-	void RenderTextNotifications();
+	void ApplyViewModelChanges(RenderParameters const & renderParameters);
+	void ApplyCanvasSizeChanges(RenderParameters const & renderParameters);	
+	void ApplyEffectiveAmbientLightIntensityChanges(RenderParameters const & renderParameters);
 
+	void RenderTextNotifications();
 	void RenderTextureNotifications();
+	void RenderHeatBlasterFlame();
+	void RenderFireExtinguisherSpray();
 
 private:
 
@@ -145,9 +239,50 @@ private:
     float mScreenToNdcX;
     float mScreenToNdcY;
 
-    float mEffectiveAmbientLightIntensity;
-
 private:
+
+	//
+	// Types
+	//
+
+#pragma pack(push, 1)
+
+	struct TextureNotificationVertex
+	{
+		vec2f vertexPositionNDC;
+		vec2f textureCoordinate;
+		float alpha;
+
+		TextureNotificationVertex(
+			vec2f const & _vertexPositionNDC,
+			vec2f const & _textureCoordinate,
+			float _alpha)
+			: vertexPositionNDC(_vertexPositionNDC)
+			, textureCoordinate(_textureCoordinate)
+			, alpha(_alpha)
+		{}
+	};
+
+	struct HeatBlasterFlameVertex
+	{
+		vec2f vertexPosition;
+		vec2f flameSpacePosition;
+
+		HeatBlasterFlameVertex()
+		{}
+	};
+
+	struct FireExtinguisherSprayVertex
+	{
+		vec2f vertexPosition;
+		vec2f spraySpacePosition;
+
+		FireExtinguisherSprayVertex()
+		{}
+	};
+
+#pragma pack(pop)
+
 
     //
     // Text machinery
@@ -263,33 +398,29 @@ private:
     std::vector<FontRenderContext> mFontRenderContexts;
 
 	//
-	// Texture machinery
+	// Texture notifications
 	//
 
 	TextureAtlasMetadata<GenericLinearTextureGroups> const & mGenericLinearTextureAtlasMetadata;
-
-#pragma pack(push, 1)
-	struct TextureNotificationVertex
-	{
-        vec2f vertexPositionNDC;
-        vec2f textureCoordinate;
-        float alpha;
-
-		TextureNotificationVertex(
-            vec2f const & _vertexPositionNDC,
-            vec2f const & _textureCoordinate,
-            float _alpha)
-			: vertexPositionNDC(_vertexPositionNDC)
-			, textureCoordinate(_textureCoordinate)
-			, alpha(_alpha)
-		{}
-	};
-#pragma pack(pop)
 
 	GameOpenGLVAO mTextureNotificationVAO;
 	std::vector<TextureNotificationVertex> mTextureNotificationVertexBuffer;
 	bool mIsTextureNotificationVertexBufferDirty;
 	GameOpenGLVBO mTextureNotificationVBO;
+
+	//
+	// Tool notifications
+	//
+
+	GameOpenGLVAO mHeatBlasterFlameVAO;
+	std::array<HeatBlasterFlameVertex, 6> mHeatBlasterFlameVertexBuffer;
+	GameOpenGLVBO mHeatBlasterFlameVBO;
+	std::optional<Render::ProgramType> mHeatBlasterFlameShaderToRender;
+
+	GameOpenGLVAO mFireExtinguisherSprayVAO;
+	std::array<FireExtinguisherSprayVertex, 6> mFireExtinguisherSprayVertexBuffer;
+	GameOpenGLVBO mFireExtinguisherSprayVBO;
+	std::optional<Render::ProgramType> mFireExtinguisherSprayShaderToRender;
 };
 
 }
